stats modified
1. health, speed, and turn-speed stats
2. Create Levels Completed Modifier
3. Round Modifier

+++ Assets/Scripts/Stats/PlayerCountStatModifier.cs
@@ -30,7 +30,7 @@ public class PlayerCountStatModifier : StatModifier
     }
 }
 
-[CustomStatModifier("Player Count Modifier", StatType.Float)]
+[CustomStatModifier("Player Count Modifier", StatType.Int)]
 public class IntPlayerCountStatModifier : StatModifier
 {
     [SerializeField]
